+++ purestorage/resource_hgroup_test.go
@@ -1,3 +1,19 @@
+/*
+   Copyright 2018 David Evans
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
 package purestorage
 
 import (
@@ -5,7 +21,7 @@ import (
 	"math/rand"
 	"testing"
 
-	"github.com/devans10/go-purestorage/flasharray"
+	"github.com/devans10/pugo/flasharray"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
 )
@@ -40,7 +56,7 @@ func TestAccResourcePureHostgroup_create_withHostlist(t *testing.T) {
 		CheckDestroy: testAccCheckPureHostgroupDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccCheckPureHostgroupConfig_withHostlist(rInt),
+				Config: testAccCheckPureHostgroupConfigWithHostlist(rInt),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckPureHostgroupExists(testAccCheckPureHostgroupResourceName, true),
 					testAccCheckPureHostgroupHosts(testAccCheckPureHostgroupResourceName, fmt.Sprintf("tfhostgrouptesthost%d", rInt), true),
@@ -59,7 +75,7 @@ func TestAccResourcePureHostgroup_create_withVolumes(t *testing.T) {
 		CheckDestroy: testAccCheckPureHostgroupDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccCheckPureHostgroupConfig_withVolumes(rInt),
+				Config: testAccCheckPureHostgroupConfigWithVolumes(rInt),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckPureHostgroupExists(testAccCheckPureHostgroupResourceName, true),
 					resource.TestCheckResourceAttr(testAccCheckPureHostgroupResourceName, "name", fmt.Sprintf("tfhostgrouptest%d", rInt)),
@@ -85,7 +101,7 @@ func TestAccResourcePureHostgroup_update(t *testing.T) {
 				),
 			},
 			{
-				Config: testAccCheckPureHostgroupConfig_rename(),
+				Config: testAccCheckPureHostgroupConfigRename(),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckPureHostgroupExists(testAccCheckPureHostgroupResourceName, true),
 					resource.TestCheckResourceAttr(testAccCheckPureHostgroupResourceName, "name", "tfhostgrouptestrename"),
@@ -110,7 +126,7 @@ func TestAccResourcePureHostgroup_update_AddandRemoveVolume(t *testing.T) {
 				),
 			},
 			{
-				Config: testAccCheckPureHostgroupConfig_withVolumes(rInt),
+				Config: testAccCheckPureHostgroupConfigWithVolumes(rInt),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckPureHostgroupExists(testAccCheckPureHostgroupResourceName, true),
 					resource.TestCheckResourceAttr(testAccCheckPureHostgroupResourceName, "name", fmt.Sprintf("tfhostgrouptest%d", rInt)),
@@ -118,7 +134,7 @@ func TestAccResourcePureHostgroup_update_AddandRemoveVolume(t *testing.T) {
 				),
 			},
 			{
-				Config: testAccCheckPureHostgroupConfig_withoutVolumes(rInt),
+				Config: testAccCheckPureHostgroupConfigWithoutVolumes(rInt),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckPureHostgroupExists(testAccCheckPureHostgroupResourceName, true),
 					resource.TestCheckResourceAttr(testAccCheckPureHostgroupResourceName, "name", fmt.Sprintf("tfhostgrouptest%d", rInt)),
@@ -140,9 +156,9 @@ func testAccCheckPureHostgroupDestroy(s *terraform.State) error {
 		_, err := client.Hostgroups.GetHostgroup(rs.Primary.ID, nil)
 		if err != nil {
 			return nil
-		} else {
-			return fmt.Errorf("hostgroup '%s' stil exists", rs.Primary.ID)
 		}
+		return fmt.Errorf("hostgroup '%s' stil exists", rs.Primary.ID)
+
 	}
 
 	return nil
@@ -194,10 +210,10 @@ func testAccCheckPureHostgroupHosts(n string, host string, exists bool) resource
 			if exists {
 				return nil
 			}
-			return fmt.Errorf("Host %s still a member of hostgroup.", host)
+			return fmt.Errorf("host %s still a member of hostgroup", host)
 		}
 		if exists {
-			return fmt.Errorf("Host %s not a member of hostgroup.", host)
+			return fmt.Errorf("host %s not a member of hostgroup", host)
 		}
 		return nil
 	}
@@ -226,11 +242,11 @@ func testAccCheckPureHostgroupVolumes(n string, volume string, exists bool) reso
 				if exists {
 					return nil
 				}
-				return fmt.Errorf("Volume %s still connected to hostgroup.", volume)
+				return fmt.Errorf("volume %s still connected to hostgroup", volume)
 			}
 		}
 		if exists {
-			return fmt.Errorf("Volume %s not connected to hostgroup.", volume)
+			return fmt.Errorf("volume %s not connected to hostgroup", volume)
 		}
 		return nil
 	}
@@ -243,7 +259,7 @@ resource "purestorage_hostgroup" "tfhostgrouptest" {
 }`, rInt)
 }
 
-func testAccCheckPureHostgroupConfig_withHostlist(rInt int) string {
+func testAccCheckPureHostgroupConfigWithHostlist(rInt int) string {
 	return fmt.Sprintf(`
 resource "purestorage_host" "tfhostgrouptesthost" {
         name = "tfhostgrouptesthost%d"
@@ -255,14 +271,14 @@ resource "purestorage_hostgroup" "tfhostgrouptest" {
 }`, rInt, rInt)
 }
 
-func testAccCheckPureHostgroupConfig_rename() string {
+func testAccCheckPureHostgroupConfigRename() string {
 	return fmt.Sprintf(`
 resource "purestorage_hostgroup" "tfhostgrouptest" {
         name = "tfhostgrouptestrename"
 }`)
 }
 
-func testAccCheckPureHostgroupConfig_withVolumes(rInt int) string {
+func testAccCheckPureHostgroupConfigWithVolumes(rInt int) string {
 	return fmt.Sprintf(`
 resource "purestorage_volume" "tfhostgrouptest-volume" {
 	name = "tfhostgrouptest-volume-%d"
@@ -278,7 +294,7 @@ resource "purestorage_hostgroup" "tfhostgrouptest" {
 }`, rInt, rInt)
 }
 
-func testAccCheckPureHostgroupConfig_withoutVolumes(rInt int) string {
+func testAccCheckPureHostgroupConfigWithoutVolumes(rInt int) string {
 	return fmt.Sprintf(`
 resource "purestorage_volume" "tfhostgrouptest-volume" {
         name = "tfhostgrouptest-volume-%d"