+++ website/terraform/dns/outputs.tf
@@ -1,0 +1,3 @@
+output "fqdn" {
+  value = "${digitalocean_record.sitename.fqdn}"
+}