+++ website/terraform/variables.tf
@@ -1,0 +1,15 @@
+variable "docker_server" {}
+
+variable "docker_email" {}
+
+variable "docker_username" {}
+
+variable "docker_password" {}
+
+variable "do_token" {}
+
+variable "sitename" {}
+
+variable "image" {}
+
+variable "cluster_name" {}